Remove constraint checking entirely from ReducedExtensionMethod
