+++ src/Compilers/CSharp/Portable/Binder/Binder_Expressions.cs
@@ -6286,20 +6286,9 @@ namespace Microsoft.CodeAnalysis.CSharp
                 // Arguments will be null if the caller is resolving to the first method group that can accept
                 // that receiver, regardless of arguments, when the signature cannot
                 // be inferred. (In the case of nameof(o.M) or the error case of o.M = null; for instance.)
-                if (analyzedArguments == null)
+                if (analyzedArguments == null && methodGroup.Methods.Count != 0)
                 {
-                    if (expression == EnclosingNameofArgument)
-                    {
-                        for (int i = methodGroup.Methods.Count - 1; i >= 0; i--)
-                        {
-                            if ((object)methodGroup.Methods[i].ReduceExtensionMethod(left.Type) == null) methodGroup.Methods.RemoveAt(i);
-                        }
-                    }
-
-                    if (methodGroup.Methods.Count != 0)
-                    {
-                        return new MethodGroupResolution(methodGroup, diagnostics.ToReadOnlyAndFree());
-                    }
+                    return new MethodGroupResolution(methodGroup, diagnostics.ToReadOnlyAndFree());
                 }
 
                 if (methodGroup.Methods.Count == 0)
