+++ src/Compilers/CSharp/Portable/Symbols/ConstraintsHelper.cs
@@ -511,6 +511,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         {
             Debug.Assert(!type.IsTupleType);
             Debug.Assert(typeArgumentsSyntax.Count == 0 /*omitted*/ || typeArgumentsSyntax.Count == type.Arity);
+            Debug.Assert(currentCompilation is object);
+
             if (!RequiresChecking(type))
             {
                 return true;
@@ -551,6 +553,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             Location location,
             DiagnosticBag diagnostics)
         {
+            Debug.Assert(currentCompilation is object);
+
             // We do not report element locations in method parameters and return types
             // so we will simply unwrap the type if it was a tuple. We are relying on
             // TypeSymbolExtensions.VisitType to dig into the "Rest" tuple so that they
@@ -817,6 +821,7 @@ hasRelatedInterfaces:
             HashSet<TypeParameterSymbol> ignoreTypeConstraintsDependentOnTypeParametersOpt)
         {
             Debug.Assert(substitution != null);
+            Debug.Assert(currentCompilation is object);
 
             // The type parameters must be original definitions of type parameters from the containing symbol.
             Debug.Assert(ReferenceEquals(typeParameter.ContainingSymbol, containingSymbol.OriginalDefinition));
@@ -873,11 +878,8 @@ hasRelatedInterfaces:
                     diagnosticsBuilder.Add(new TypeParameterDiagnosticInfo(typeParameter, new CSDiagnosticInfo(ErrorCode.ERR_UnmanagedConstraintNotSatisfied, containingSymbol.ConstructedFrom(), typeParameter, typeArgument.Type)));
                     return false;
                 }
-                else if (managedKind == ManagedKind.UnmanagedWithGenerics && currentCompilation is object)
+                else if (managedKind == ManagedKind.UnmanagedWithGenerics)
                 {
-                    // In batch scenarios currentCompilation should always be non-null, but the public API
-                    // can check constraints in multiple places and there may not be a current compilation.
-                    // For these scenarios, we'll skip checking
                     var csDiagnosticInfo = MessageID.IDS_FeatureUnmanagedConstructedTypes
                         .GetFeatureAvailabilityDiagnosticInfoOpt((CSharpCompilation)currentCompilation);
                     if (csDiagnosticInfo != null)
