+++ src/Compilers/CSharp/Portable/Symbols/ReducedExtensionMethodSymbol.cs
@@ -7,6 +7,7 @@ using System.Globalization;
 using System.Threading;
 using Microsoft.CodeAnalysis.CSharp.Symbols;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
 
@@ -29,7 +30,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
         /// is applicable, and satisfies type parameter constraints, based on the
         /// "this" argument type. Otherwise, returns null.
         /// </summary>
-        public static MethodSymbol Create(MethodSymbol method, TypeSymbol receiverType, Compilation compilation)
+        public static MethodSymbol Create(MethodSymbol method, TypeSymbol receiverType)
         {
             Debug.Assert(method.IsExtensionMethod && method.MethodKind != MethodKind.ReducedExtension);
             Debug.Assert(method.ParameterCount > 0);
@@ -37,7 +38,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
             HashSet<DiagnosticInfo> useSiteDiagnostics = null;
 
-            method = method.InferExtensionMethodTypeArguments(receiverType, compilation, ref useSiteDiagnostics);
+            method = InferExtensionMethodTypeArguments(method, receiverType, ref useSiteDiagnostics);
             if ((object)method == null)
             {
                 return null;
@@ -95,6 +96,118 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             _typeMap = TypeMap.Empty.WithAlphaRename(reducedFrom, this, out _typeParameters);
             _typeArguments = _typeMap.SubstituteTypes(reducedFrom.TypeArgumentsWithAnnotations);
         }
+
+        /// <summary>
+        /// If the extension method is applicable based on the "this" argument type, return
+        /// the method constructed with the inferred type arguments. If the method is not an
+        /// unconstructed generic method, type inference is skipped. If the method is not
+        /// applicable, or if constraints when inferring type parameters from the "this" type
+        /// are not satisfied, the return value is null.
+        /// </summary>
+        private static MethodSymbol InferExtensionMethodTypeArguments(MethodSymbol method, TypeSymbol thisType, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            Debug.Assert(method.IsExtensionMethod);
+            Debug.Assert((object)thisType != null);
+
+            if (!method.IsGenericMethod || method != method.ConstructedFrom)
+            {
+                return method;
+            }
+
+            // We never resolve extension methods on a dynamic receiver.
+            if (thisType.IsDynamic())
+            {
+                return null;
+            }
+
+            var containingAssembly = method.ContainingAssembly;
+            var errorNamespace = containingAssembly.GlobalNamespace;
+            var conversions = new TypeConversions(containingAssembly.CorLibrary);
+
+            // There is absolutely no plausible syntax/tree that we could use for these
+            // synthesized literals.  We could be speculatively binding a call to a PE method.
+            var syntaxTree = CSharpSyntaxTree.Dummy;
+            var syntax = (CSharpSyntaxNode)syntaxTree.GetRoot();
+
+            // Create an argument value for the "this" argument of specific type,
+            // and pass the same bad argument value for all other arguments.
+            var thisArgumentValue = new BoundLiteral(syntax, ConstantValue.Bad, thisType) { WasCompilerGenerated = true };
+            var otherArgumentType = new ExtendedErrorTypeSymbol(errorNamespace, name: string.Empty, arity: 0, errorInfo: null, unreported: false);
+            var otherArgumentValue = new BoundLiteral(syntax, ConstantValue.Bad, otherArgumentType) { WasCompilerGenerated = true };
+
+            var paramCount = method.ParameterCount;
+            var arguments = new BoundExpression[paramCount];
+
+            for (int i = 0; i < paramCount; i++)
+            {
+                var argument = (i == 0) ? thisArgumentValue : otherArgumentValue;
+                arguments[i] = argument;
+            }
+
+            var typeArgs = MethodTypeInferrer.InferTypeArgumentsFromFirstArgument(
+                conversions,
+                method,
+                arguments.AsImmutable(),
+                useSiteDiagnostics: ref useSiteDiagnostics);
+
+            if (typeArgs.IsDefault)
+            {
+                return null;
+            }
+
+            int firstNullInTypeArgs = -1;
+
+            // For the purpose of constraint checks we use error type symbol in place of type arguments that we couldn't infer from the first argument.
+            // This prevents constraint checking from failing for corresponding type parameters. 
+            var notInferredTypeParameters = PooledHashSet<TypeParameterSymbol>.GetInstance();
+            var typeParams = method.TypeParameters;
+            var typeArgsForConstraintsCheck = typeArgs;
+            for (int i = 0; i < typeArgsForConstraintsCheck.Length; i++)
+            {
+                if (!typeArgsForConstraintsCheck[i].HasType)
+                {
+                    firstNullInTypeArgs = i;
+                    var builder = ArrayBuilder<TypeWithAnnotations>.GetInstance();
+                    builder.AddRange(typeArgsForConstraintsCheck, firstNullInTypeArgs);
+
+                    for (; i < typeArgsForConstraintsCheck.Length; i++)
+                    {
+                        var typeArg = typeArgsForConstraintsCheck[i];
+                        if (!typeArg.HasType)
+                        {
+                            notInferredTypeParameters.Add(typeParams[i]);
+                            builder.Add(TypeWithAnnotations.Create(ErrorTypeSymbol.UnknownResultType));
+                        }
+                        else
+                        {
+                            builder.Add(typeArg);
+                        }
+                    }
+
+                    typeArgsForConstraintsCheck = builder.ToImmutableAndFree();
+                    break;
+                }
+            }
+
+            // For the purpose of construction we use original type parameters in place of type arguments that we couldn't infer from the first argument.
+            var typeArgsForConstruct = typeArgs;
+            if (firstNullInTypeArgs != -1)
+            {
+                var builder = ArrayBuilder<TypeWithAnnotations>.GetInstance();
+                builder.AddRange(typeArgs, firstNullInTypeArgs);
+
+                for (int i = firstNullInTypeArgs; i < typeArgsForConstruct.Length; i++)
+                {
+                    var typeArgForConstruct = typeArgsForConstruct[i];
+                    builder.Add(typeArgForConstruct.HasType ? typeArgForConstruct : TypeWithAnnotations.Create(typeParams[i]));
+                }
+
+                typeArgsForConstruct = builder.ToImmutableAndFree();
+            }
+
+            return method.Construct(typeArgsForConstruct);
+        }
+
 
         internal override MethodSymbol CallsiteReducedFromMethod
         {
